requirements.txt
# Core dependencies
fastapi>=0.104.0
openai>=1.0.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
pyyaml>=6.0

# Streamlit UI
streamlit>=1.28.0

# LLM Providers (optional - install if using real LLM mode)
openai>=1.0.0
anthropic>=0.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
ruff>=0.1.0
